Move DocsSummarizer to query_helpers
