--- ols/src/docs/__init__.py
@@ -1,1 +1,0 @@
-"""Documentation context handlings."""
