+++ ols/src/query_helpers/docs_summarizer.py
