+++ tests/mock_classes/mock_llama_index.py
@@ -8,13 +8,13 @@ class MockLlamaIndex:
 
     Example usage in a test:
 
-        @patch("ols.src.docs.docs_summarizer.load_index_from_storage", new=MockLlamaIndex)
+        @patch("ols.src.query_helpers.docs_summarizer.load_index_from_storage", new=MockLlamaIndex)
         def test_xyz():
 
         or within test function or test method:
-        with patch("ols.src.docs.docs_summarizer.load_index_from_storage", new=MockLlamaIndex):
+        with patch("ols.src.query_helpers.docs_summarizer.load_index_from_storage", new=MockLlamaIndex):
             some test steps
-    """
+    """  # noqa E501
 
     def __init__(self, *args, **kwargs):
         """Store all provided arguments for later usage."""
