+++ tests/unit/docs/test_doc_summarizer.py
@@ -3,16 +3,18 @@
 import os
 from unittest.mock import patch
 
-from ols.src.docs.docs_summarizer import DocsSummarizer
+from ols.src.query_helpers.docs_summarizer import DocsSummarizer
 from ols.utils import config
 from tests.mock_classes.llm_loader import mock_llm_loader
 from tests.mock_classes.mock_llama_index import MockLlamaIndex
 
 
-@patch("ols.src.docs.docs_summarizer.LLMLoader", new=mock_llm_loader(None))
-@patch("ols.src.docs.docs_summarizer.ServiceContext.from_defaults")
-@patch("ols.src.docs.docs_summarizer.StorageContext.from_defaults")
-@patch("ols.src.docs.docs_summarizer.load_index_from_storage", new=MockLlamaIndex)
+@patch("ols.src.query_helpers.docs_summarizer.LLMLoader", new=mock_llm_loader(None))
+@patch("ols.src.query_helpers.docs_summarizer.ServiceContext.from_defaults")
+@patch("ols.src.query_helpers.docs_summarizer.StorageContext.from_defaults")
+@patch(
+    "ols.src.query_helpers.docs_summarizer.load_index_from_storage", new=MockLlamaIndex
+)
 @patch.dict(os.environ, {"TEI_SERVER_URL": "localhost"}, clear=True)
 def test_summarize(storage_context, service_context):
     """Basic test for DocsSummarizer using mocked index and query engine."""
